Implementation of two algorithms to find the fibonacci number of n.
File fibonacci.py has the implementation of two algorithms to find the
fibonacci number of n, one is recursive and the other one uses lists. With
the tests that have been writen in the file test_fibonacci.py is clear that
the fib_list function is faster than fib_recurs function.